feat: 공간인덱스 일부 적용
일반검색 공간인덱스 적용

+++ Backend/ssaeng-go-jip/src/main/java/com/ssaenggojip/property/service/PropertyService.java
@@ -42,7 +42,7 @@ public class PropertyService {
 
     private SearchProperty mapToDto(Property p) {
         return SearchProperty.builder()
-                .id(p.getId().intValue())
+                .id(p.getId())
                 .dealType(p.getDealType())
                 .price(p.getPrice())
                 .rentPrice(p.getRentPrice())
